backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart
pydantic>=2.6.0
huggingface_hub>=0.20.0
openai>=1.0.0
python-dotenv>=1.0.0




